Refactor: SortDropDown 컴포넌트 삭제 -> Main 으로 재이동
