--- src/components/SortDropDown/SortDropDown.module.scss
@@ -1,6 +1,0 @@
-.dropdownContainer {
-  width: 100%;
-  display: flex;
-  justify-content: right;
-  margin: 10px 0px;
-}
