+++ src/pages/Main/Main.module.scss
@@ -5,3 +5,10 @@
   justify-content: center;
   align-items: center;
 }
+
+.dropdownContainer {
+  width: 100%;
+  display: flex;
+  justify-content: right;
+  margin: 10px 0px;
+}
